Unify UTCTimeField
Replace the UTCTimeField occurences (one in ppi_geotag.py, one in dhcp6
and one in tls/handshake) by a single one
